ML-27 User refactor
