+++ src/users/viewsets/user/serializer.py
@@ -6,7 +6,7 @@ from users.models import User
 class UserSerializer(serializers.ModelSerializer):
     class Meta:
         model = User
-        fields = ('uuid', 'username', 'name', 'lastname', 'email', 'password')
+        fields = ('uuid', 'username', 'name', 'lastname', 'email', 'password', 'role')
         extra_kwargs = {
             'password': {'write_only': True}
         }
